+++ lib/shared/value_objects/single_value.dart
@@ -1,0 +1,13 @@
+import 'package:equatable/equatable.dart';
+
+/// For classes that holds a single value.
+abstract class SingleValue<T> with EquatableMixin {
+  const SingleValue(this.value);
+
+  /// The value.
+  final T value;
+
+  /// Equality by [value].
+  @override
+  List<Object?> get props => [value];
+}